+++ src/components/transfer/RegularTransferTable.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next';
 import { Button, Table } from 'reactstrap';
-import AccountDropdown from './AccountDropdown';
+import AccountDropdown from './table/AccountDropdown';
 import { CSSProperties, useState } from 'react';
 import { deleteRegular, insertRegular, updateRegular } from '@/lib/regularReq';
